refactor(console): split ApiKeysManager monolith into focused modules
Break the ~1100-line ApiKeysManager.tsx into a clean components/api-keys/
module with single-responsibility files:

- constants.ts        storage keys, shared class strings, copy, reset options
- format.ts           pure presentation helpers (usd/time/status/usage color)
- limits.ts           form-input -> sparse PATCH value parsers
- Modal.tsx           shared dialog shell
- KeyForm.tsx         create/edit form
- SecretReveal.tsx    show-once secret reveal
- ConfirmBody.tsx     confirm dialog body
- KeyCard.tsx         key row (+ chip/usage-bar/icon-action internals)
- useApiKeys.ts       data + mutations + console-key bookkeeping hook
- ApiKeysManager.tsx  thin orchestrator (modal state only)
- index.ts            barrel export

No behavior change. Page import updated to @/components/api-keys.
eslint clean, vitest 162 pass, next build succeeds.
