+++ console-ui/src/components/api-keys/Modal.tsx
@@ -1,0 +1,51 @@
+"use client";
+
+import { useEffect } from "react";
+import { X } from "lucide-react";
+
+// Modal is the shared dialog shell: backdrop, Escape-to-close, sticky header.
+export function Modal({
+  open,
+  onClose,
+  title,
+  children,
+  maxWidth = "max-w-md",
+}: {
+  open: boolean;
+  onClose: () => void;
+  title: string;
+  children: React.ReactNode;
+  maxWidth?: string;
+}) {
+  useEffect(() => {
+    if (!open) return;
+    const handler = (e: KeyboardEvent) => {
+      if (e.key === "Escape") onClose();
+    };
+    document.addEventListener("keydown", handler);
+    return () => document.removeEventListener("keydown", handler);
+  }, [open, onClose]);
+
+  if (!open) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center p-3 sm:p-4">
+      <div className="absolute inset-0 bg-ink/40 backdrop-blur-sm" onClick={onClose} />
+      <div
+        className={`relative w-full ${maxWidth} max-h-[88vh] overflow-y-auto rounded-2xl bg-bg-white border border-border-dim shadow-xl`}
+      >
+        <div className="sticky top-0 z-10 flex items-center justify-between bg-bg-white px-5 py-4 border-b border-border-dim">
+          <h3 className="text-base font-semibold text-text-primary">{title}</h3>
+          <button
+            onClick={onClose}
+            className="p-1.5 rounded-lg hover:bg-bg-hover text-text-tertiary transition-colors"
+            aria-label="Close"
+          >
+            <X size={16} />
+          </button>
+        </div>
+        <div className="px-5 py-5">{children}</div>
+      </div>
+    </div>
+  );
+}
