+++ console-ui/src/components/api-keys/index.ts
@@ -1,0 +1,1 @@
+export { ApiKeysManager } from "./ApiKeysManager";
